Refactors the images service
This change just moves code around and introduces some helper
methods in an attempt to make our code a little more readable.

The majority of the complexity has been moved into an image filter
that we can apply to an image.

This also introduces some basic tests for the routes just as a
quick check to see if anything is broken.